almost all of dolphin matched (#2036)
* TRK full match

* remove trk asm

* ar done

* cleanup some dolphin headers

* more dolphin cleanup

* cleanup / GD fully matched

* almost all of GX fully matched

* GX / Mtx full matched

* most of OS done

* pad done

* most of VI

* remove asm

* forgot couple vec funcs

* couple JUtility matches

+++ include/dolphin/os/OSThread.h
@@ -68,7 +68,7 @@ struct OSThread {
     OSMutexQueue owned_mutexes;
     OSThreadLink active_threads_link;
     u8* stack_base;
-    u8* stack_end;
+    u32* stack_end;
     u8* error_code;
     void* data[2];
 };
